Fixed issues with EF Core tables and added data generation code using Bogus

+++ src/Data.Core/Model/Customer.cs
@@ -23,7 +23,7 @@ namespace Data.Core.Model
         public string City { get; set; }
         
         [ForeignKey("RegionId"), Required]
-        public long RegionId { get; init; }
+        public int RegionId { get; init; }
         public Region Region { get; init; }
         
         [Column("ZipCode"), Required]
